refactor: 출석 확인 기능 CrewAttendances 객체로 이동
