+++ src/main/java/attendance/dto/ConfirmAttendanceDto.java
@@ -1,0 +1,5 @@
+package attendance.dto;
+
+import attendance.domain.Attendance;
+
+public record ConfirmAttendanceDto(Attendance attendance) {}
